外部キーデータ型変更

+++ database/migrations/2021_12_11_125123_create_recipes_table.php
@@ -18,7 +18,7 @@ class CreateRecipesTable extends Migration
             $table->integer('user_id');
             $table->date('cooking_day');
             $table->string('recipe_time');
-            $table->increments('menu_id')->unsigned();
+            $table->bigInteger('menu_id')->unsigned();
             $table->timestamps();
         });
     }
